% Auto-generated: do not edit by hand
\name{treeDashComponent}

\alias{treeDashComponent}

\title{TreeDashComponent component}

\description{

}

\usage{
treeDashComponent(id=NULL, symbol=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{symbol}{Named list. }
}
